+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 NOTE: Subcrates have their own changelogs: [bevy-tnua-physics-integration-layer](physics-integration-layer/CHANGELOG.md), [bevy-tnua-rapier](rapier3d/CHANGELOG.md), [bevy-tnua-xpbd](xpbd3d/CHANGELOG.md).
 
 ## [Unreleased]
+### Fixed
+- Use boost for stopping in `TnuaBuiltinWalk` (fixes
+  https://github.com/idanarye/bevy-tnua/issues/39)
 
 ## 0.14.1 - 2024-01-14
 ### Fixed